SellOrders

* feat: create sell orders

* feat: create buy orders

* fix: favorite nfts and create not setting id and sell 500 error

* fix: profile with new schema and added new tab inventory

+++ interfaces/src/main/java/ar/edu/itba/paw/service/MailingService.java
@@ -2,10 +2,12 @@ package ar.edu.itba.paw.service;
 
 import org.springframework.scheduling.annotation.Async;
 
+import java.math.BigDecimal;
+
 public interface MailingService {
 
     @Async
-    void sendOfferMail(String buyerEmail, String sellerMail, String nftName, String nftAddress, float nftPrice);
+    void sendOfferMail(String buyerEmail, String sellerMail, String nftName, String nftAddress, BigDecimal nftPrice);
 
     @Async
     void sendRegisterMail(String userEmail, String username);
